Only use Cloudfront for static content
Fronting the entire site with Cloudfront is problematic for image uploads in new blog posts. So skip CF for the site itself and just use an APIGW domain. Instead, we now have a CF distro specifically for static content, plus supporting changes in WP to use the CF URL for stylesheets and such.